Working on metrics tracking

+++ Systems/SimulationSystems.cs
@@ -14,18 +14,78 @@ public class SimulationSystems
     public static void Clear()
     {
         List<int> carComponents = ComponentManager.GetComponent<Car>();
-            for (int i = carComponents.Count - 1; i >= 0; i--)
-            {
-                EntityManager.RemoveEntity(carComponents[i]);
-            }
-            
-            List<int> pathSegmentComponents = ComponentManager.GetComponent<PathSegment>();
-            for (int i = pathSegmentComponents.Count - 1; i >= 0; i--)
-            {
-                EntityManager.RemoveEntity(pathSegmentComponents[i]);
-            }
+        for (int i = carComponents.Count - 1; i >= 0; i--)
+        {
+            EntityManager.RemoveEntity(carComponents[i]);
+        }
+        
+        List<int> pathSegmentComponents = ComponentManager.GetComponent<PathSegment>();
+        for (int i = pathSegmentComponents.Count - 1; i >= 0; i--)
+        {
+            EntityManager.RemoveEntity(pathSegmentComponents[i]);
+        }
 
-            RoadMesh roadMesh = EntityManager.GetGlobalComponent<RoadMesh>();
-            roadMesh.DestroyAll();
+        RoadMesh roadMesh = EntityManager.GetGlobalComponent<RoadMesh>();
+        roadMesh.DestroyAll();
+        
+        EntityManager.AddComponentToGlobalEntity(new SimulationSuper());
+    }
+
+    public static void RunSimulation(float virutalDt)
+    {
+        SimulationSuper simSuper = EntityManager.GetGlobalComponent<SimulationSuper>();
+
+        simSuper.CongestionTimer += virutalDt;
+        simSuper.SimTimer += virutalDt;
+
+        if (simSuper.CongestionTimer > simSuper.CONGESTION_TIMER_MAX)
+        {
+            CaptureCurrentAverageCongestion();
+        }
+        
+        if (simSuper.SimTimer < simSuper.MaxSimTime ||  simSuper.SimCount > simSuper.MaxSimCount)
+            return;
+
+        simSuper.SimTimer = 0;
+        simSuper.SimCount += 1;
+        simSuper.GenBehaviorDistribution += simSuper.DistributionChangeValue;
+        
+        //create new simulation metrics object and fill lists
+        
+        //clear cars next
+        List<int> carComponents = ComponentManager.GetComponent<Car>();
+        for (int i = carComponents.Count - 1; i >= 0; i--)
+        {
+            EntityManager.RemoveEntity(carComponents[i]);
+        }
+        
+        //gen new cars with same config
+        CarSystems.GenerateCars(simSuper.GenCarCount, simSuper.GenSeed, simSuper.GenBehaviorDistribution);
+    }
+
+    public static void CaptureCurrentAverageCongestion()
+    {
+        SimulationSuper simSuper = EntityManager.GetGlobalComponent<SimulationSuper>();
+        List<int> segmentEntities = ComponentManager.GetComponent<PathSegment>();
+
+        float totalCongestion = 0;
+        
+        foreach (int segmentEntity in segmentEntities)
+        {
+            PathSegment segment = ComponentManager.GetEntityComponent<PathSegment>(segmentEntity);
+            totalCongestion += segment.CongestionCost;
+        }
+        
+        simSuper.CurrentSimTracking.AverageCongestionCollection.Add((totalCongestion / segmentEntities.Count, simSuper.SimTimer));
+    }
+
+    public static void IncrementDestinations(bool isReroute)
+    {
+        SimulationSuper simSuper = EntityManager.GetGlobalComponent<SimulationSuper>();
+
+        if (isReroute)
+            simSuper.CurrentSimTracking.RerouteDestinations += 1;
+        else
+            simSuper.CurrentSimTracking.BasicDestinations += 1;
     }
 }
